update quiz react code
