+++ app/javascript/packs/quiz.js
@@ -1,0 +1,1 @@
+const render = require("../src/quiz/render.jsx");
