--- app/javascript/packs/quiz/index.js
@@ -1,1 +1,0 @@
-const render = require("./render.jsx");
