+++ app/javascript/src/gameColourWords/components/gameTimer.jsx
