+++ app/javascript/src/gameColourWords/components/score.jsx
