+++ app/javascript/src/gameColourWords/components/scoreList.jsx
