+++ app/javascript/src/gameColourWords/components/strike.jsx
