+++ app/javascript/src/gameColourWords/components/strikeList.jsx
