+++ app/javascript/src/gameColourWords/index.js
