+++ app/javascript/src/gameColourWords/stores/game_store.js
