+++ app/javascript/src/gameColourWords/stores/result_store.js
