+++ app/javascript/src/quiz/render.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import Quiz from './components/quiz.jsx'
+
+document.addEventListener('DOMContentLoaded', () => {
+  const node = document.getElementById('renderer')
+  const props = JSON.parse(node.getAttribute('data-props'))
+  ReactDOM.render(
+    <Quiz {...props} />,
+    node
+  )
+})
+
+
